Reorg: Move dtos to subpackage
