+++ src/main/java/com/ginsberg/gatherers4j/dto/WithCount.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Todd Ginsberg
+ * Copyright 2025 Todd Ginsberg
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.ginsberg.gatherers4j;
+package com.ginsberg.gatherers4j.dto;
 
 import org.jspecify.annotations.Nullable;
 
-public record WithIndex<TYPE>(
-        long index,
-        @Nullable TYPE value
+public record WithCount<VALUE>(
+        @Nullable VALUE value,
+        long count
 ) {
 }
