+++ src/test/java/com/ginsberg/gatherers4j/AccumulatingGathererTest.java
@@ -16,6 +16,7 @@
 
 package com.ginsberg.gatherers4j;
 
+import com.ginsberg.gatherers4j.dto.WithIndex;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
