+++ src/test/java/com/ginsberg/gatherers4j/CrossGathererTest.java
@@ -16,6 +16,7 @@
 
 package com.ginsberg.gatherers4j;
 
+import com.ginsberg.gatherers4j.dto.Pair;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
